+++ test/safe_transformer_test.js
@@ -17,30 +17,20 @@ describe('SafeTransformer', function() {
     let script = 'test/fixtures/sample.php';
     let transformer = new SafeTransformer(new Minifier('php'));
 
-    it('should remove the inline comments', done => {
-      transformer.transform(script).subscribe(output => {
-        /* eslint-disable quotes */
-        expect(output).to.contain("<?= 'Hello World!' ?>");
-        /* eslint-enable quotes */
-      }, done, done);
+    it('should remove the inline comments', async () => {
+      expect(await transformer.transform(script)).to.contain("<?= 'Hello World!' ?>");
     });
 
-    it('should remove the multi-line comments', done => {
-      transformer.transform(script).subscribe(output => {
-        expect(output).to.contain('namespace dummy; class Dummy');
-      }, done, done);
+    it('should remove the multi-line comments', async () => {
+      expect(await transformer.transform(script)).to.contain('namespace dummy; class Dummy');
     });
 
-    it('should remove the single-line comments', done => {
-      transformer.transform(script).subscribe(output => {
-        expect(output).to.contain('$className = get_class($this); return $className;');
-      }, done, done);
+    it('should remove the single-line comments', async () => {
+      expect(await transformer.transform(script)).to.contain('$className = get_class($this); return $className;');
     });
 
-    it('should remove the whitespace', done => {
-      transformer.transform(script).subscribe(output => {
-        expect(output).to.contain('__construct() { }');
-      }, done, done);
+    it('should remove the whitespace', async () => {
+      expect(await transformer.transform(script)).to.contain('__construct() { }');
     });
   });
 });